Bluetooth support working
It seems when the bluetooth is plugged in, whether it's active or not, my temperatures start to read funny.
If it's unplugged, whether the code thinks it's there or not, the temperatures are OK.

I suspect it is drawing too much current for my RAMPS13 along with the LCD and such.